+++ Package.swift
@@ -14,14 +14,14 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/shibapm/Logger", from: "0.1.0"),
-        .package(url: "https://github.com/nerdishbynature/octokit.swift", from: "0.9.0"),
         .package(url: "https://github.com/mxcl/Version", from: "1.0.0"),
+        .package(url: "https://github.com/nerdishbynature/octokit.swift", from: "0.10.0"),
         // Danger Plugins
         // Dev dependencies
         .package(url: "https://github.com/shibapm/Komondor", from: "1.0.0"), // dev
         .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.35.8"), // dev
         .package(url: "https://github.com/Realm/SwiftLint", from: "0.28.1"), // dev
-        .package(url: "https://github.com/pointfreeco/swift-snapshot-testing.git", from: "1.0.0"), // dev
+        .package(url: "https://github.com/pointfreeco/swift-snapshot-testing.git", .branch("master")), // dev
         .package(url: "https://github.com/shibapm/Rocket", from: "0.4.0"), // dev
         .package(url: "https://github.com/jpsim/Yams.git", from: "1.0.0"), // dev
         .package(url: "https://github.com/eneko/SourceDocs", from: "0.5.1"), // dev