+++ Tests/DangerTests/XCTestManifests.swift
@@ -1,3 +1,4 @@
+
 #if !canImport(ObjectiveC)
     import XCTest